requirements.txt
atlassian-python-api
beautifulsoup4
langchain
langchain-core
langchain-openai
langchain-community
openai
tiktoken
chromadb
cohere
python-dotenv